A ton of changes
split up all initialization functions so that now injection wf generation and running wf generation is the same.
Now a lot of functions are unused such as get_pe and injectWaveform

Fixed bug so that when skyloc but not time is getting updated the waveforms are still projected correctly

Also untangled most of initialization code!
In doing so moved parameters from kwargs directly into LnLikelihoodManager. Now we should be able to run on real data with any waveform so long as the h5 parameters correspond to that waveform.

TODO is to just pass a reference waveform or to make the spins work for the given waveform

We can run on real data with --Tcut-seconds set, but what might not work is --Tcut-cycles (because the h5 file I have should correspond to an aligned spin waveform which i simply do not have unfortunately)